qt: Add UI debug mode (-debug-ui)
If this mode is enabled the content of the css files will become pushed to the UI every second. It also forces some UI elements to show up which are actually only shown under special conditions (e.g. watch balance labels). Its required to set a custom css directory with -custom-css-dir to enable this.