qt: Runtime theme changes and improved debug-ui
This commit enables runtime theme changes. Means no more client
restart required if the theme gets changed in the options dialog.

The stylesheet update process of the ui debug mode gets also improved by this
changes. Prior the stylesheets were pushed to the UI every second no matter
if there was a change or not. This is changing to only pushing them to the
UI if there were any changes made to the stylesheets.